feat-app: 여행 계획 페이지 헤더 구현
