+++ application/lib/feature/travel_read/model/travel_visit_model.dart
@@ -1,5 +1,3 @@
-import 'package:application_new/feature/travel_read/model/travel_visit_rating_model.dart';
-import 'package:application_new/feature/travel_read/model/travel_visit_reason.dart';
 import 'package:application_new/shared/model/image_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
@@ -8,7 +6,6 @@ part 'travel_visit_model.g.dart';
 
 @freezed
 class TravelVisitModel with _$TravelVisitModel {
-
   const factory TravelVisitModel({
     required int id,
     required int seq,
@@ -21,4 +18,30 @@ class TravelVisitModel with _$TravelVisitModel {
 
   factory TravelVisitModel.fromJson(Map<String, dynamic> json) =>
       _$TravelVisitModelFromJson(json);
+}
+
+@freezed
+class TravelVisitRatingModel with _$TravelVisitRatingModel {
+  const factory TravelVisitRatingModel({
+    required int satisfaction,
+    required int revisit,
+    required int recommend,
+  }) = _TravelVisitRatingModel;
+
+  factory TravelVisitRatingModel.fromJson(Map<String, dynamic> json) =>
+      _$TravelVisitRatingModelFromJson(json);
+}
+
+enum TravelVisitReason {
+  landmark,
+  goodReviews,
+  media,
+  goodExperience,
+  recommendation,
+  easyAccess,
+  goodAmenities,
+  goodValue,
+  unexpectedFind,
+  education,
+  other
 }
