--- application/lib/feature/travel_read/model/travel_visit_reason.dart
@@ -1,14 +1,0 @@
-
-enum TravelVisitReason {
-  landmark,
-  goodReviews,
-  media,
-  goodExperience,
-  recommendation,
-  easyAccess,
-  goodAmenities,
-  goodValue,
-  unexpectedFind,
-  education,
-  other
-}
